requirements.txt
openai>=0.28
streamlit-feedback
azure-identity==1.15.0
azure-ai-documentintelligence==1.0.0b3
streamlit>=1.28
streamlit-chat
asyncio
pymupdf
tiktoken
requests
pandas
biopython
pyautogen
semantic-kernel
matplotlib
seaborn
